+++ backend/src/controllers/admin/glossary.js
@@ -2,11 +2,11 @@ const useController = require("../useController/index.js");
 const what = { single: "glossary", plural: "glossary" };
 const controller = useController(null, what);
 const path = require("path");
+const fs = require("fs");
 
 /**
- * Get all glossary terms. Uses fastify-markdown plugin to parse glossary markdown into HTML.
+ * Get all glossary terms.
  *
- * @see https://github.com/freezestudio/fastify-markdown
  * @param   {FastifyRequest} request FastifyRequest is an instance of the standard http or http2 request objects.
  * @param   {FastifyReply}   reply   FastifyReply is an instance of the standard http or http2 reply types.
  * @returns {object}
@@ -15,9 +15,11 @@ const getAll = async (request, reply) => {
   controller.userRequires(request, "PMO-Manager-Edit-Capability", reply);
 
   try {
-    const result = await reply.markdown(
-      path.join(__dirname, "..", "..", "..", "docs", "Glossary", "Glossary.md")
+    const stream = fs.createReadStream(
+      path.join(__dirname, "..", "..", "..", "docs", "Glossary", "Glossary.html"),
+      "utf8"
     );
+    const result = await reply.header("Content-Type", "application/octet-stream").send(stream);
     return result;
   } catch (err) {
     return controller.failedQuery(reply, err, what);